Clean up integration
